+++ internal/components/components.go
@@ -46,6 +46,7 @@ import (
 	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/stackdriverexporter"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/sumologicexporter"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/tanzuobservabilityexporter"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/zipkinexporter"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/extension/bearertokenauthextension"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/extension/fluentbitextension"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/extension/httpforwarder"
@@ -110,13 +111,17 @@ func Components() (component.Factories, error) {
 
 	delete(factories.Exporters, "kafka")
 	delete(factories.Exporters, "opencensus")
+	delete(factories.Exporters, "zipkin")
+
 	delete(factories.Extensions, "oidc")
 	delete(factories.Extensions, "bearertokenauth")
+
 	delete(factories.Processors, "probabilistic_sampler")
 	delete(factories.Processors, "attributes")
 	delete(factories.Processors, "span")
 	delete(factories.Processors, "resource")
 	delete(factories.Processors, "filter")
+
 	delete(factories.Receivers, "opencensus")
 	delete(factories.Receivers, "kafka")
 	delete(factories.Receivers, "hostmetrics")
@@ -213,6 +218,7 @@ func Components() (component.Factories, error) {
 		stackdriverexporter.NewFactory(),
 		sumologicexporter.NewFactory(),
 		tanzuobservabilityexporter.NewFactory(),
+		zipkinexporter.NewFactory(),
 	}
 	for _, exp := range factories.Exporters {
 		exporters = append(exporters, exp)
